+++ packages/player/ci.config.js
