+++ packages/player/package.json
