+++ packages/player/playwright.config.js
