+++ packages/player/setup/global.ts
