+++ packages/player/tsconfig.json
